requirements.txt
asgiref
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
environs
python-dotenv
django-crispy-forms
crispy-bootstrap5
sounddevice 
pydub
soundfile
openai
spacy
pygame